ugc_api/pyproject.toml
[tool.poetry]
name = "ugc_api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
orjson = "^3.7.2"
pydantic = "^1.9.1"
aiokafka = "^0.7.2"
kafka-python = "^2.0.2"
PyJWT = "^2.4.0"
gunicorn = "^20.1.0"
uvicorn = "^0.17.6"
backoff = "^2.1.2"
requests = "^2.28.0"
motor = "^3.0.0"
sentry-sdk = "^1.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
